A-HIGH-NNN.check_env_improvements
