+++ docs/feature_stories/FS_29_54_67_86.dir_structure.md
@@ -85,20 +85,81 @@ For example:
 *   When the project is installed to environment `target_env_a`, `@/conf/` should be a symlink to `@/dst/target_env_a`.
 *   When the project is installed to environment `target_env_b`, `@/conf/` should be a symlink to `@/dst/target_env_b`.
 
+## `@/conf/python_env.conf.bash`
+
+This config script is used by `@/exe/bootstrap_env.bash` to `source`:
+*   path to initial Python interpreter (to init `venv` with)
+*   path to `venv` to activate
+
+If `venv` is wrongly initialized:
+*   update `@/conf/python_env.conf.bash`
+*   delete old `venv`
+*   initialize new `venv` by re-running `@/exe/bootstrap_env.bash`
+
+## `@/conf/env_packages.txt`
+
+It contains `pip freeze` output (AKA `requirements.txt`, just renamed) with Python packages and their versions.
+
+Script `@/exe/bootstrap_env.bash` installs packages according to this file.
+
+## `@/conf/argrelay_client.json`
+
+Config for `argrelay` client.
+
+Unlike `argrelay` server YAML config, config for `argrelay` client is JSON for performance reasons.
+
+## `@/conf/argrelay_server.yaml`
+
+Config for `argrelay` server.
+
+## `@/conf/argrelay_plugin.yaml`
+
+Config for `argrelay` plugin(s).
+
 ## `@/exe/bootstrap_env.bash`
+
+It bootstraps environment according to configs in `@/conf/`.
 
 See FS_85_33_46_53 `bootstrap_env`.
 
 ## `@/exe/check_env.bash`
 
+It checks environment for any issues.
+
 See FS_36_17_84_44 `check_env`.
 
 ## `@/exe/dev_shell.bash`
+
+It starts fully configured shell according to configs in `@/conf/` to try `argrelay`-linked commands.
+
+To avoid starting it, see `@/exe/shell_env.bash`.
 
 See FS_58_61_77_69 `dev_shell`.
 
 ## `@/conf/dev_shell_env.bash`
 
-It is used by `@/exe/dev_shell.bash`.
+It is a hook used to additionally configure `@/exe/dev_shell.bash`.
+
+## `@/exe/run_argrelay_server`
+
+It runs `argrelay` server (in foreground).
+
+## `@/exe/run_argrelay_client`
+
+It is **not** used directly - instead, Bash calls it on `Alt+Shift+Q`-query or `Tab`-completion.
+
+## `@/exe/shell_env.bash`
+
+It is what `~/.bashrc` has to `source` to avoid starting `@/exe/dev_shell.bash`.
+
+## `@/conf/shell_env.conf.bash`
+
+It lists `argrelay`-linked commands (to be stored in `@/bin` and exposed via `PATH`).
+
+## `@/var/argrelay_client.server_index`
+
+It contains index of one of the configured servers client used for the previous request.
+
+See FS_93_18_57_91 client fail over to redundant servers.
 
 [FS_16_07_78_84.conf_dir_priority.md]: FS_16_07_78_84.conf_dir_priority.md
