+++ docs/feature_stories/FS_36_17_84_44.check_env.md
@@ -17,8 +17,6 @@ This script may overlap with verifications done by FS_85_33_46_53 `bootstrap_env
 And, similar to `bootstrap_env`, `check_env` is installed to target integration project as a copy<br/>
 so that it can be run on project clone before bootstrapping the environment.
 
-See also FS_86_73_43_45 server control scripts.
-
 # Idea
 
 The set of checks to be performed should be open for extension by the target project.
