+++ docs/feature_stories/FS_61_67_08_53.non_searchable_arg_types.md
@@ -9,6 +9,9 @@ Some arg types are non-searchable: there is no way to index their arg types as `
 For example, if there is a function to download a file,
 remote file path is such an arg (arbitrary, not part of indexed data).
 
+TODO: Figure out how to propagate extra Bash completion parameters to enable, for example,
+      auto-completion of files and directories (done locally by Bash instead of `argrelay`).
+
 TODO: Figure out how to specify interrogation step for these arg types.
       At the moment, such args are provided without interrogation spec - user can simply add extra args
       to the command line. They will be sent to client on invocation and plugin client side can try
