+++ docs/feature_stories/FS_70_55_40_99.splitting_config_files.md
@@ -15,3 +15,5 @@ SomePlugin.default:
     plugin_config:
         !include "dst/env_name/plugin_config/SomePlugin.default.yaml"
 ```
+
+All references to files are relative to `@/` = `argrelay_dir` - see FS_29_54_67_86 `dir_structure`.
