+++ docs/feature_stories/FS_86_73_43_45.server_control_scripts.md
@@ -5,6 +5,7 @@ feature_status: TODO
 ---
 
 TODO: Re-write the scripts in Python (for greater test-ability).
+      Currently, they are implemented as part of `@/exe/argrelay_common_lib.bash`.
 
 This feature provides standard framework-supported server control scripts:
 *   start
