+++ docs/feature_stories/FS_93_18_57_91.client_fail_over_to_redundant_servers.md
@@ -8,7 +8,7 @@ keywords: client-side fail over, redundant servers
 
 This feature is for client-side fail-over to other configured redundant servers.
 
-It uses special state file `@/var/argrelay_client.server_index`
+It uses `@/var/argrelay_client.server_index` state file
 to select one of the several configured server connection.
 
 If connection fails, it selects (round-robin) next connection.
