+++ docs/task_refs/TODO_86_57_50_38.config_only_common_delegator_behavior.md
@@ -1,8 +1,16 @@
 
-TODO: TODO_86_57_50_38: config-only for common delegator behavior
+TODO: TODO_86_57_50_38: (config-only for) common delegator behavior
 
-For example, if delegator cannot work with non-signled-out `data_envelop` and needs to fail with proper error
-(rather than exception), provide configuration for such behaviors.
+Actually, this applies to behavior of any delegator (config-only or not)
+because config-only uses same common APIs just being able to be configured.
+
+For example:
+
+*   If delegator cannot work with non-singled-out `data_envelop` and needs to fail with proper error
+    (rather than exception), provide configuration for such behaviors.
+
+*   If delegator (or specific funct) is supposed to fail on unconsumed/unrecognized/remaining arg,
+    it this can be part of common support by the framework.
 
 This makes some of the behavior of FS_49_96_50_77 config-only plugins available for any plugins.
