+++ src/argrelay/check_env/PluginCheckEnvServerIndex.py
@@ -18,7 +18,7 @@ _server_index_field_name = "server_index"
 
 class PluginCheckEnvServerIndex(PluginCheckEnvAbstract):
     """
-    Simple plugin to report server index for FS_93_18_57_91 client fail over.
+    This plugin reports `server_index` for FS_93_18_57_91 client fail over.
     """
 
     # noinspection PyMethodMayBeStatic
