+++ src/argrelay/custom_integ/value_constants.py
@@ -1,12 +1,12 @@
 """
 """
 
-func_id_goto_host = "func_id_goto_host"
+func_id_goto_host_ = "func_id_goto_host"
 func_id_goto_service_ = "func_id_goto_service"
 func_id_goto_git_repo_ = "func_id_goto_git_repo"
 
-func_id_desc_host = "func_id_desc_host"
-func_id_desc_service = "func_id_desc_service"
+func_id_desc_host_ = "func_id_desc_host"
+func_id_desc_service_ = "func_id_desc_service"
 
 func_id_list_host_ = "func_id_list_host"
 func_id_list_service_ = "func_id_list_service"
